+++ cherryd/openflow/of13/action.go
@@ -19,13 +19,13 @@ type Action struct {
 	openflow.BaseAction
 }
 
-func marshalOutput(p uint32) ([]byte, error) {
+func marshalOutput(p uint) ([]byte, error) {
 	v := make([]byte, 16)
 	binary.BigEndian.PutUint16(v[0:2], uint16(OFPAT_OUTPUT))
 	binary.BigEndian.PutUint16(v[2:4], 16)
 
 	var port uint32
-	switch uint(p) {
+	switch p {
 	case openflow.PortTable:
 		port = OFPP_TABLE
 	case openflow.PortAll:
@@ -35,7 +35,7 @@ func marshalOutput(p uint32) ([]byte, error) {
 	case openflow.PortAny:
 		port = OFPP_ANY
 	default:
-		port = p
+		port = uint32(p)
 	}
 	binary.BigEndian.PutUint32(v[4:8], port)
 	// We don't support buffer ID and partial PACKET_IN
@@ -87,7 +87,7 @@ func (r *Action) MarshalBinary() ([]byte, error) {
 	}
 
 	if ok, port := r.Output(); ok {
-		v, err := marshalOutput(uint32(port))
+		v, err := marshalOutput(port)
 		if err != nil {
 			return nil, err
 		}
